[ICU-3222] fix compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/14007 Former-commit-id: fb91824fc4a0b7ceb3b57fe1dfad1af2868231fe

+++ icu4c/source/tools/makeconv/gencnvex.c
@@ -272,7 +272,7 @@ CnvExtWrite(NewConverter *cnvData, const UConverterStaticData *staticData,
 #endif
 
     if(VERBOSE) {
-        printf("size of extension data: %ld\n", top);
+        printf("size of extension data: %ld\n", (long)top);
     }
 
     /* return the number of bytes that should have been written */
